pyproject.toml
[project]
name = "BugViperion"
version = "0.1.0"
description = "Code Review Tool - Repository Ingestion Engine with Tree-sitter and Neo4j"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Code Review Team", email = "team@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    # Database
    "neo4j>=6.0.3",
    # Environment
    "python-dotenv>=1.2.1",
    # Tree-sitter core and language parsers
    "tree-sitter>=0.21.0",
    "tree-sitter-language-pack>=0.6.0",
    # Config file parsing
    "pyyaml>=6.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    # File handling and patterns
    "pathspec>=0.12.0",
    "nbformat>=5.9.0",
    "nbconvert>=7.16.0",
    # API dependencies
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "pygithub>=2.8.1",
    # Agent dependencies
    "pydantic-ai[openrouter]>=1.47.0",
    "openai>=2.15.0",
    "logfire>=4.19.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0", 
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/Pavel401/BugViper"
Repository = "https://github.com/Pavel401/BugViper"
Issues = "https://github.com/Pavel401/BugViper/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "db", "agents"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py313"]
